+++ operator/CHANGELOG.md
@@ -17,6 +17,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 - Webhook and AIStore type validation ensures consistency between `spec.auth.profileRef` and `spec.configToUpdate.auth.enabled`.
 - `AIStoreAuth`
   - Support for `spec.deployment.pod.annotations` to set custom annotations on the AuthN Deployment pod template.
+  - Webhook validation rejects specs when neither `spec.adminSecret` nor `spec.deployment.pod.annotations` are set.
 
 ### Changed